% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annInner.R
\name{annInner}
\alias{annInner}
\title{ANN model in inner CV loop}
\usage{
annInner(
  xTrain,
  yTrain,
  xVal,
  yVal,
  DA,
  fitness,
  method,
  nodes,
  threshold,
  stepmax
)
}
\arguments{
\item{xTrain}{Training data (samples as rows; variables as columns)}

\item{yTrain}{Training response}

\item{xVal}{Validation data}

\item{yVal}{Validation response (for tuning)}

\item{DA}{Logical for discriminant analysis (classification)}

\item{fitness}{Fitness function ('MISS', 'AUROC' or 'RMSEP')}

\item{method}{method for ann}

\item{nodes}{node}

\item{threshold}{the Threshold of step size}

\item{stepmax}{Maximum step in each annInner loop}

\item{layers}{Number of layers and how many node in each layed. If multiple layers, it writes as c(n,n,n)}
}
\value{
An object containing: (`miss`, `auc` or `rmsep`) A fitness metric &`virank` variable importance rankings
}
\description{
ANN model in inner CV loop
}
