% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annpredict.R
\name{annpredict}
\alias{annpredict}
\title{ANN predictions for outer segments and consensus models}
\usage{
annpredict(xTrain, yTrain, xTest, yTest, DA, nodes, threshold, stepmax, method)
}
\arguments{
\item{xTrain}{xTrain}

\item{yTrain}{yTrain}

\item{xTest}{xTest}

\item{yTest}{yTest}

\item{DA}{DA}

\item{nodes}{nodes}

\item{threshold}{threshold}

\item{stepmax}{stepmax}

\item{method}{method}
}
\value{
The predicted value of yTest
}
\description{
ANN predictions for outer segments and consensus models
}
