% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{checkinput}
\alias{checkinput}
\title{Check input}
\usage{
checkinput(
  X,
  Y,
  ML,
  DA,
  method,
  fitness,
  nInner,
  nOuter,
  varRatio,
  scale,
  modReturn,
  logg,
  parallel
)
}
\arguments{
\item{X}{The original data of X, not the result after onehotencoding}

\item{Y}{The original data of Y}

\item{ML}{ML in MUVR2}

\item{DA}{DA in MUVR2}

\item{method}{RF or PLS so far in MUVR2}

\item{fitness}{fitness in MUVR2}

\item{nInner}{nInnerin MUVR2}

\item{nOuter}{nOuter in MUVR2}

\item{varRatio}{varRatio in MUVR2}

\item{scale}{scale}

\item{modReturn}{modReturn in MUVR2}

\item{logg}{logg in MUVR2}

\item{parallel}{parallel in MUVR2}
}
\value{
correct_input: the original input(call) and the real input used in MUVR2 when you enter your input
}
\description{
This can be run to test if the command input of parameters contradict each
other and check the structure of the data. If something goes wrong, warning
messages are given.
}
\examples{
data("freelive2")
checkinput(X = XRVIP2,
           Y = YR2,  ## YR2 a numeric variable
           DA = FALSE,
           fitness="RMSEP")
}
