% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Make a confusion matrix from a MUVR object}
\usage{
confusionMatrix(MVObj, model = "mid")
}
\arguments{
\item{MVObj}{A MUVR object (classification analysis)}

\item{model}{min, mid or max model}
}
\value{
A confusion matrix of actual vs predicted class
}
\description{
Make a confusion matrix from a MUVR object
}
