% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMUVR.R
\name{convertMUVR}
\alias{convertMUVR}
\title{The original result MUVR model has the class of 'MVObject'.
This function is to transform its classname from 'MVObject' to 'MUVR'
This function also transform item of the output in MUVR():
VIP-->VIRank
VIPPerRep-->VIRankPerRep}
\usage{
convertMUVR(MUVRclassObject)
}
\arguments{
\item{MUVRclassObject}{A 'MUVR' class object}
}
\value{
A 'MUVR' class Object where all is transformed to
}
\description{
The original result MUVR model has the class of 'MVObject'.
This function is to transform its classname from 'MVObject' to 'MUVR'
This function also transform item of the output in MUVR():
VIP-->VIRank
VIPPerRep-->VIRankPerRep
}
