% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_samesequence.R
\name{factor_samesequence}
\alias{factor_samesequence}
\title{factor_samesequence: This function is to change the order of items in the levels of a factor variable}
\usage{
factor_samesequence(X, level, sequence)
}
\arguments{
\item{X}{a factor variable}

\item{level}{levels if want to manually set}

\item{sequence}{default as F, if T normal factor}
}
\value{
X_factor
}
\description{
factor_samesequence: This function is to change the order of items in the levels of a factor variable
}
