% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldVector.R
\name{foldVector}
\alias{foldVector}
\title{Convert folds in listformat to vectorformat. Specifcy for each ID which group they are in}
\usage{
foldVector(foldList, ID)
}
\arguments{
\item{foldList}{fold list from vectSamp or uniqDASamp}

\item{ID}{Vector of sampling unit identifier}
}
\value{
a vector with folds
}
\description{
Convert folds in listformat to vectorformat. Specifcy for each ID which group they are in
}
\examples{
Y <- rep(LETTERS[1:2],10)
ID <- 1:20
folds <- 5
foldList <- uniqDASamp(Y, ID, folds)
foldVect <- foldVector(foldList, ID)
}
