% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBER.R
\name{getBER}
\alias{getBER}
\title{Get BER}
\usage{
getBER(actual, predicted, weigh_added = FALSE, weighing_matrix)
}
\arguments{
\item{actual}{Vector of actual classifications of samples}

\item{predicted}{Vector of predicted classifications of samples}

\item{weigh_added}{To add a weighing matrix when it is classification}

\item{weighing_matrix}{The matrix used to get a misclassification score}
}
\value{
BER
}
\description{
Get Balanced Error Rate (BER) in classification.
}
\examples{
data("mosquito")
actual <- Yotu
predicted <- sampling_from_distribution(actual)
getBER(actual, predicted)
}
