% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMISS.R
\name{getMISS}
\alias{getMISS}
\title{Get number of misclassifications}
\usage{
getMISS(actual, predicted, weigh_added = FALSE, weighing_matrix)
}
\arguments{
\item{actual}{Vector of actual classifications of samples}

\item{predicted}{Vector of predicted classifications of samples}

\item{weigh_added}{Boolean, add a weighing matrix when it is classification}

\item{weighing_matrix}{The matrix used to get a misclassification score}
}
\value{
number of misclassifications
}
\description{
Get number of misclassifications from classification analysis.
}
\examples{
data("mosquito")
actual <- Yotu
predicted <- sampling_from_distribution(actual)
getMISS(actual, predicted)
}
