% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPred.R
\name{ggplotPred}
\alias{ggplotPred}
\title{Plot predictions for PLS regression (ggplot2)}
\usage{
ggplotPred(Ytrue, Ypreds)
}
\arguments{
\item{Ytrue}{True value of Y, should be a vector}

\item{Ypreds}{Predicted value of Y; a vector, or a matrix/data frame with one
row per observation}
}
\value{
A \code{ggplot} object
}
\description{
\code{ggplot2} version of \code{\link[=plotPred]{plotPred()}}. Grey points are the individual predictions
(one series per column of \code{Ypreds}); black points are their consensus (row
means).
}
\examples{
\donttest{
data("freelive2")
Ypreds <- sampling_from_distribution(YR2)
ggplotPred(YR2, Ypreds)
}
}
\seealso{
\code{\link[=plotPred]{plotPred()}} for the base graphics version
}
