% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUVR2-package.R
\docType{package}
\name{MUVR2-package}
\alias{MUVR2-package}
\title{MUVR2: Multivariate Methods with Unbiased Variable Selection}
\description{
Predictive multivariate modelling for metabolomics. Types: classification and regression. Methods: partial least squares, random forest and elastic net. Data structures: paired and unpaired. Validation: repeated double cross-validation (Westerhuis et al. (2008)\doi{10.1007/s11306-007-0099-6}, Filzmoser et al. (2009)\doi{10.1002/cem.1225}). Variable selection: performed internally, through tuning in the inner cross-validation loop.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MetaboComp/MUVR2}
  \item \url{https://metabocomp.github.io/MUVR2/}
  \item Report bugs at \url{https://github.com/MetaboComp/MUVR2/issues}
}

}
\author{
\strong{Maintainer}: Yingxiao Yan \email{yingxiao@chalmers.se}

Authors:
\itemize{
  \item Yingxiao Yan \email{yingxiao@chalmers.se}
  \item Carl Brunius \email{carl.brunius@chalmers.se}
}

}
\keyword{internal}
