% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls}
\alias{pls}
\title{PLS regression
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)}
\usage{
pls(
  x,
  y,
  ncomp = 2,
  max.iter = 500,
  tol = 1e-06,
  near.zero.var = TRUE,
  scale = TRUE
)
}
\arguments{
\item{x}{Numeric matrix of predictors with the rows as individual observations. missing values (NAs) are allowed.}

\item{y}{Numeric matrix of response(s) with the rows as individual observations matching X. missing values (NAs) are allowed.}

\item{ncomp}{Positive Integer. The number of components to include in the model. Default to 2.}

\item{max.iter}{Integer, the maximum number of iterations. Default to 100.}

\item{tol}{Positive numeric used as convergence criteria/tolerance during the iterative process. Default to 1e-06.}

\item{near.zero.var}{near_zero_variance in MUVR}

\item{scale}{scale or not}
}
\value{
a plsMUVR object
}
\description{
PLS regression
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
\examples{
\dontrun{
data(freelive2)
pls(XRVIP2,
YR2)

}
}
