% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda}
\alias{plsda}
\title{PLS-DA
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)}
\usage{
plsda(
  x,
  y,
  ncomp = 2,
  max.iter = 500,
  tol = 1e-06,
  near.zero.var = TRUE,
  scale = TRUE
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{ncomp}{ncomp}

\item{max.iter}{max.iter}

\item{tol}{tol}

\item{near.zero.var}{near.zero.var}

\item{scale}{scale}
}
\value{
a plsdaMUVR object
}
\description{
PLS-DA
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
\examples{
\dontrun{
data("mosquito")
MUVR2::plsda(Xotu,
      Yotu)
}
}
