% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.plsdaMUVR}
\alias{predict.plsdaMUVR}
\title{Predict plsda
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)}
\usage{
predict.plsdaMUVR(object, newdata, scale = TRUE, onlyPred = FALSE, ...)
}
\arguments{
\item{object}{a plsdaMUVR object}

\item{newdata}{new data}

\item{scale}{scale in MUVR}

\item{onlyPred}{Boolean for whether to report back predictions only (defaults to FALSE)}

\item{...}{other arguments}
}
\value{
plsda predictions
}
\description{
Predict plsda
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
