% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.plsMUVR}
\alias{predict.plsMUVR}
\title{Predict pls
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)}
\usage{
\method{predict}{plsMUVR}(object, newdata, onlyPred = FALSE, scale = TRUE, ...)
}
\arguments{
\item{object}{a plsMUVR object}

\item{newdata}{new data}

\item{onlyPred}{Boolean for whether to report back predictions only (defaults to FALSE)}

\item{scale}{scale in MUVR}

\item{...}{other argument}
}
\value{
pls prediction
}
\description{
Predict pls
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
\examples{
\dontrun{
data(freelive2)
object<-MUVR2::pls(XRVIP2,YR2)
predict(object,XRVIP2)
}
}
