% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_calculation.R
\name{Q2_calculation}
\alias{Q2_calculation}
\title{Q2 calculation}
\usage{
Q2_calculation(yhat, y)
}
\arguments{
\item{yhat}{prediction values}

\item{y}{real values}
}
\value{
Q2
}
\description{
Q2 calculation
}
