% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfInner.R
\name{rfInner}
\alias{rfInner}
\title{RF model in inner CV loop}
\usage{
rfInner(
  xTrain,
  yTrain,
  xVal,
  yVal,
  DA,
  fitness,
  ntree,
  mtry,
  method,
  weigh_added,
  weighing_matrix
)
}
\arguments{
\item{xTrain}{Training data (samples as rows; variables as columns)}

\item{yTrain}{Training response}

\item{xVal}{Validation data}

\item{yVal}{Validation response (for tuning)}

\item{DA}{Logical for discriminant analysis (classification)}

\item{fitness}{Fitness function ('MISS', 'AUROC' or 'RMSEP')}

\item{ntree}{See original function (`randomForest`). Passed from wrapper.}

\item{mtry}{See original function (`randomForest`). Passed from wrapper.}

\item{method}{Choice of Random Forest implementation (randomForest, ranger or Rborist). Passed from wrapper.}

\item{weigh_added}{To add a weighing matrix when it is classfication}

\item{weighing_matrix}{The matrix used for get a miss classfication score}
}
\value{
An object containing:

(`miss`, `auc` or `rmsep`) A fitness metric

`virank` variable importance rankings
}
\description{
RF model in inner CV loop
}
