% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfPred_test_rfkeep.R
\name{rfPred_test_rfkeep}
\alias{rfPred_test_rfkeep}
\title{RF predictions for outer segments and consensus models test version, the original version is saved in rfPred}
\usage{
rfPred_test_rfkeep(
  xTrain,
  yTrain,
  xTest,
  yTest,
  ntree = 500,
  DA,
  keep.forest = FALSE,
  method,
  keep_variables = NULL
)
}
\arguments{
\item{xTrain}{xTrain}

\item{yTrain}{yTrain}

\item{xTest}{xTest}

\item{yTest}{yTest}

\item{ntree}{ntree}

\item{DA}{DA}

\item{keep.forest}{keep.forest}

\item{method}{method}

\item{keep_variables}{The variables to be kept}
}
\value{
The predicted value of yTest
}
\description{
RF predictions for outer segments and consensus models test version, the original version is saved in rfPred
}
