% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_from_distribution.R
\name{sampling_from_distribution}
\alias{sampling_from_distribution}
\title{Sampling from the distribution of something}
\usage{
sampling_from_distribution(X, upperlimit, lowerlimit, extend, n)
}
\arguments{
\item{X}{a vector (numeric or factor) where the distribution/probabality will be generated}

\item{upperlimit}{if X is numeric, set upper limit}

\item{lowerlimit}{if X is numeric, set lower limit}

\item{extend}{If X is numeric, how much you want to extend from the lower and upper existing X.}

\item{n}{How many you want to sample}
}
\value{
a resampled thing
}
\description{
Sampling from the distribution of something
}
