% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmInner.R
\name{svmInner}
\alias{svmInner}
\title{Support vector machine model in inner CV loop}
\usage{
svmInner(xTrain, yTrain, xVal, yVal, DA, fitness, kernel, method)
}
\arguments{
\item{xTrain}{Training data (samples as rows; variables as columns)}

\item{yTrain}{Training response}

\item{xVal}{Validation data}

\item{yVal}{Validation response (f´´or tuning)}

\item{DA}{Logical for discriminant analysis (classification)}

\item{fitness}{Fitness function ('MISS', 'AUROC' or 'RMSEP')}

\item{kernel}{What kernel to choose: polynomial, radical, linear, sigmoid}

\item{method}{Choice of Support vector machine implementation ("svm", "ksvm", "svmlight"). Passed from wrapper.}
}
\value{
An object containing:(`miss`,`ber`,`auc` or `rmsep`) A fitness metric and `virank` variable importance rankings
}
\description{
Support vector machine model in inner CV loop
}
