% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmpredict.R
\name{svmpredict}
\alias{svmpredict}
\title{RF predictions for outer segments and consensus models}
\usage{
svmpredict(
  xTrain,
  yTrain,
  xTest,
  yTest,
  method = c("svm", "ksvm", "svmlight"),
  kernel = c("vanilladot", "polydot", "rbfdot"),
  DA
)
}
\arguments{
\item{xTrain}{xTrain}

\item{yTrain}{yTrain}

\item{xTest}{xTest}

\item{yTest}{yTest}

\item{method}{method}

\item{kernel}{The build in method of svm "linear","radical","polynomial","sigmoid"}

\item{DA}{DA or not, coming form MUVR}
}
\value{
The predicted value of yTest
}
\description{
RF predictions for outer segments and consensus models
}
