% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqDASamp.R
\name{uniqDASamp}
\alias{uniqDASamp}
\title{Stratified sampling for classification and unique individuals
Even if it is stratified sampling, it may lead to some fold have more than 1 samples than others}
\usage{
uniqDASamp(Y, ID, folds)
}
\arguments{
\item{Y}{Classes (Std Y variable)}

\item{ID}{Individual identifier (unique)}

\item{folds}{Number of folds}
}
\value{
A master list of segments
}
\description{
Stratified sampling for classification and unique individuals
Even if it is stratified sampling, it may lead to some fold have more than 1 samples than others
}
\examples{
\dontrun{
data(mosquito)
Y <- Yotu
ID <- 1:length(Yotu)
folds <- 5
uniqDASamp(Y, ID, folds)
}
}
