% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectSamp.R
\name{vectSamp}
\alias{vectSamp}
\title{Sampling of a vector
Random sampling of a vector into n groups
Bug: Sampling is faulty when length(vect) < n}
\usage{
vectSamp(vect, n = 4)
}
\arguments{
\item{vect}{A vector to be sampled into groups}

\item{n}{Number of groups}
}
\value{
a list with n groups containing sub sampled `vect`
}
\description{
Sampling of a vector
Random sampling of a vector into n groups
Bug: Sampling is faulty when length(vect) < n
}
\examples{
\dontrun{
data("mosquito")
vectSamp(Yotu)
}
}
