% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{vip}
\alias{vip}
\title{Extract PLS VIP values
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)}
\usage{
vip(object)
}
\arguments{
\item{object}{pls(da)MUVR object}
}
\value{
vip object
}
\description{
Extract PLS VIP values
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
\examples{
\dontrun{
data("freelive2")
object <- MUVR2::pls(XRVIP2,YR2)
vip_object<-vip(object)
vip_object
}
}
