% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YR}
\alias{YR}
\title{Rye consumption for the rye metabolomics regression tutorial}
\format{
A numeric vector of length 112.
}
\usage{
data(freelive)
}
\description{
Wholegrain rye intake for the 112 samples in \link{XRVIP}.
}
\seealso{
\link{XRVIP}, \link{IDR}
}
\keyword{data}
