% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YR2}
\alias{YR2}
\title{Rye consumption for the rye metabolomics regression tutorial, using unique individuals}
\format{
A numeric vector of length 58.
}
\usage{
data(freelive2)
}
\description{
Wholegrain rye intake for the 58 individuals in \link{XRVIP2}. This is the continuous
target variable of the regression examples.
}
\seealso{
\link{XRVIP2}, \link{IDR2}
}
\keyword{data}
