% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Make a confusion matrix from a MUVR object}
\usage{
confusionMatrix(MVObj, model = "mid")
}
\arguments{
\item{MVObj}{A MUVR object (classification analysis)}

\item{model}{min, mid or max model}
}
\value{
A confusion matrix of actual vs predicted class
}
\description{
Make a confusion matrix from a MUVR object
}
\examples{
nRep <- 2 # Number of MUVR2 repetitions
nOuter <- 3 # Number of outer cross-validation segments
varRatio <- 0.75 # Proportion of variables kept per iteration
method <- 'RF' # Selected core modeling algorithm
classModel <- MUVR2(
  X = Xotu,
  Y = Yotu,
  nRep = nRep,
  nOuter = nOuter,
  varRatio = varRatio,
  method = method
)
confusionMatrix(classModel)

}
