% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmsep.R
\name{get_rmsep}
\alias{get_rmsep}
\title{get rmsep for regression}
\usage{
get_rmsep(actual, predicted)
}
\arguments{
\item{actual}{Vector of actual classifications of samples}

\item{predicted}{Vector of predicted classifications of samples}
}
\value{
rmsep
}
\description{
get rmsep for regression
}
\examples{
data("mosquito")
actual <- YR2
predicted <- sampling_from_distribution(actual)
get_rmsep(actual, predicted)
}
