% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVar.R
\name{getVar}
\alias{getVar}
\title{The function will obtain the min, mid, max number of variables for an object generated from the rdCVnet() function}
\usage{
getVar(
  rdCVnetObject,
  option = c("quantile", "fitness"),
  fit_curve = c("loess", "gam"),
  span = 1,
  k = 5,
  outlier = c("none", "IQR", "residual"),
  robust = 0.05,
  quantile = 0.25
)
}
\arguments{
\item{rdCVnetObject}{an object obtained from the rdCVnet() function}

\item{option}{quantile or fitness: In which way do you perform variable selection}

\item{fit_curve}{gam or loess method for fitting the curve in the fitness option}

\item{span}{parameter for using loess to fit curve in the fitness option:  how smooth the curve need to be}

\item{k}{parameter for using gam to fit curve in the fitness option}

\item{outlier}{if remove ourlier variables or not. There are 3 options: "none","IRQ", "residual"}

\item{robust}{If the option is fitness, robust parameter decides how much deviation it is allowed from the optimal perdiction performance for min and max variabel selection.}

\item{quantile}{If the option is quantile, this value decides the cut for the first quantile, ranging from 0 to 0.5.}
}
\value{
a rdCVnet object
}
\description{
The function will obtain the min, mid, max number of variables for an object generated from the rdCVnet() function
}
\examples{
\dontrun{

}
}
