% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVIRank.R
\name{getVIRank}
\alias{getVIRank}
\title{Extract autoselected variables from MUVR model object}
\usage{
getVIRank(MUVRclassObject, model = "mid", n, all = FALSE)
}
\arguments{
\item{MUVRclassObject}{An object of MUVR class}

\item{model}{Which model to use ("min", "mid" (default), or "max")}

\item{n}{customize values}

\item{all}{logical, to get the ranks of all variable or not}
}
\value{
Data frame with order, name and average rank of variables (`order`, `name` & `rank`)
}
\description{
Extract autoselected variables from MUVR model object
}
