% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbiplotPLS.R
\name{ggbiplotPLS}
\alias{ggbiplotPLS}
\title{PLS biplot (ggplot2)}
\usage{
ggbiplotPLS(
  fit,
  comps = 1:2,
  xCol,
  labPlSc = TRUE,
  labs,
  vars,
  labPlLo = TRUE,
  colSc = "black",
  colLo = "red",
  supLeg = FALSE
)
}
\arguments{
\item{fit}{A PLS fit (e.g. from \code{MUVRclassObject$Fit[[2]]})}

\item{comps}{Which two components to plot}

\item{xCol}{(Optional) Continuous vector for a colour gradient over the
observations (e.g. the Y vector in a regression analysis)}

\item{labPlSc}{Boolean to plot observation (sample) names (defaults to TRUE)}

\item{labs}{(Optional) Label names for the observations}

\item{vars}{Which variables to plot (names in \code{rownames(loadings)})}

\item{labPlLo}{Boolean to plot variable names (defaults to TRUE)}

\item{colSc}{Colour for observation scores (only used if \code{xCol} is omitted)}

\item{colLo}{Colour for variable loadings (defaults to red)}

\item{supLeg}{Boolean for whether to suppress the legend}
}
\value{
A \code{ggplot} object
}
\description{
\code{ggplot2} version of \code{\link[=biplotPLS]{biplotPLS()}}. Plots observation scores and variable
loadings of a PLS fit in one set of axes: scores on the primary (bottom/left)
axes, loadings on the secondary (top/right) axes.
}
\examples{
\donttest{
data("freelive2")
regrModel <- MUVR2(X = XRVIP2,
                   Y = YR2,
                   nRep = 2,
                   nOuter = 3,
                   method = "PLS",
                   modReturn = TRUE)
ggbiplotPLS(regrModel$Fit[[2]],
            comps = 1:2,
            xCol = YR2,
            labPlSc = FALSE,
            labPlLo = FALSE)
}
}
\seealso{
\code{\link[=biplotPLS]{biplotPLS()}} for the base graphics version
}
