% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPCA.R
\name{ggplotPCA}
\alias{ggplotPCA}
\title{PCA score plot (ggplot2)}
\usage{
ggplotPCA(pca, PC1 = 1, PC2 = 2, colVar, symbVar, main = NULL)
}
\arguments{
\item{pca}{A \code{prcomp} object}

\item{PC1}{Principal component on x-axis}

\item{PC2}{Principal component on y-axis}

\item{colVar}{Continuous variable for colouring observations}

\item{symbVar}{Categorical/discrete variable for multiple plot symbols}

\item{main}{Optional plot title}
}
\value{
A \code{ggplot} object
}
\description{
\code{ggplot2} version of \code{\link[=plotPCA]{plotPCA()}}. Plots PCA scores with the possibility to
choose which principal components to show, to colour observations by a
continuous variable and to use different plotting symbols for a categorical
variable.
}
\details{
Unlike \code{\link[=plotPCA]{plotPCA()}}, this function has no \code{file} argument: it returns a
\code{ggplot} object, which you save with \code{ggplot2::ggsave()}.
}
\examples{
data("freelive2")
pca_object <- prcomp(XRVIP2)
ggplotPCA(pca_object, colVar = YR2)
}
\seealso{
\code{\link[=plotPCA]{plotPCA()}} for the base graphics version
}
