% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H0_reference.R
\name{H0_reference}
\alias{H0_reference}
\title{Get reference distribution for resampling tests}
\usage{
H0_reference(Y, n = 1000, fitness = c("Q2", "BER", "MISS", "AUROC"), ...)
}
\arguments{
\item{Y}{the target variable}

\item{n}{number of permutations to run}

\item{fitness}{number of repetitions for each permutation (defaults to value of actual model)}

\item{...}{additional arguments for sampling from distribution}
}
\value{
a histogram of reference distribution
}
\description{
Make reference distribution for resampling tests to assess overfitting.
}
\examples{
\donttest{
data("freelive2")
H0_reference(YR2)
}
}
