% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IDR}
\alias{IDR}
\title{Subject identifiers for the rye metabolomics regression tutorial}
\usage{
data(freelive)
}
\description{
Subject identifiers for the rye metabolomics regression tutorial
}
\keyword{data}
