% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoding.R
\name{onehotencoding}
\alias{onehotencoding}
\title{One_hot_encoding
Each factor and character variable with n categories(>2) will be transformed to n variables
Each factor and character variable with 2 categories will be transformed to one 01 numeric dummy  variable
Each factor and character variable with 1 categories will be transformed to one numeric variable that only has value 1
Each factor and character variable with 0 categories will be transformed to one numeric variable that only has value -999
Each logical variable will be transformed to one 01 numeric dummy variable}
\usage{
onehotencoding(X)
}
\arguments{
\item{X}{Data frame. This data frame could contain numeric, factor, character and logical variable}
}
\value{
new_X_matrix:The new matrix that transform all variables to numeric variables
}
\description{
One_hot_encoding
Each factor and character variable with n categories(>2) will be transformed to n variables
Each factor and character variable with 2 categories will be transformed to one 01 numeric dummy  variable
Each factor and character variable with 1 categories will be transformed to one numeric variable that only has value 1
Each factor and character variable with 0 categories will be transformed to one numeric variable that only has value -999
Each logical variable will be transformed to one 01 numeric dummy variable
}
