% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPred.R
\name{plotPred}
\alias{plotPred}
\title{Plot predictions
At present, this function only supports predictions for PLS regression type problems}
\usage{
plotPred(Ytrue, Ypreds)
}
\arguments{
\item{Ytrue}{True value of Y}

\item{Ypreds}{Predicted value of Y}
}
\description{
Plot predictions
At present, this function only supports predictions for PLS regression type problems
}
\examples{
\dontrun{
data("freelive2")
Ytrue<-YR2
Ypreds<-sampling_from_distribution(YR2)
plotPred(Ytrue,Ypreds)
}
}
