% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVAL.R
\name{plotVAL}
\alias{plotVAL}
\title{Plot validation metric
Produces a plot of validation metric vs number of variables in model (inner segment)}
\usage{
plotVAL(MUVRclassObject, show_outlier = T)
}
\arguments{
\item{MUVRclassObject}{An object of class `MUVR`}

\item{show_outlier}{deciding show outlier on the information of not}
}
\value{
A plot
}
\description{
Plot validation metric
Produces a plot of validation metric vs number of variables in model (inner segment)
}
