% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predMV.R
\name{predMV}
\alias{predMV}
\title{Predict MV object using a MUVR class object and a X testing set
At present, this function only supports predictions for PLS regression type problems}
\usage{
predMV(MUVRclassobject, newdata, model = "min")
}
\arguments{
\item{MUVRclassobject}{An 'MUVR' class object obtained from the MVWrap function}

\item{newdata}{New data for which to predict outcomes}

\item{model}{What type of model to plot ('min', 'mid' or 'max'). Defaults to 'mid'.}
}
\value{
A pdf with plots of results from multivariate predictions
}
\description{
Predict MV object using a MUVR class object and a X testing set
At present, this function only supports predictions for PLS regression type problems
}
\examples{
\dontrun{

}
}
