% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predMV.R
\name{predMV}
\alias{predMV}
\title{Predict MV object using a MUVR class object and a X testing set
At present, this function only supports predictions for PLS regression type problems}
\usage{
predMV(MUVRclassobject, newdata, model = "min")
}
\arguments{
\item{MUVRclassobject}{An 'MUVR' class object obtained from the MVWrap function}

\item{newdata}{New data for which to predict outcomes}

\item{model}{What type of model to plot ('min', 'mid' or 'max'). Defaults to 'mid'.}
}
\value{
A pdf with plots of results from multivariate predictions
}
\description{
Predict MV object using a MUVR class object and a X testing set
At present, this function only supports predictions for PLS regression type problems
}
\examples{
\dontrun{
data("freelive2")
nRep <- 2
nOuter <- 4
varRatio <-0.6
regrModel <- MUVR2(X = XRVIP2,
                   Y = YR2,
                   nRep = nRep,
                   nOuter = nOuter,
                   varRatio = varRatio,
                   method = "PLS",
                   modReturn=T)
predMV(regrModel,XRVIP2)
}
}
