% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_calculation_forclassfication.R
\name{Q2_calculation_forclassification}
\alias{Q2_calculation_forclassification}
\title{Q2 calculation for classification}
\usage{
Q2_calculation_forclassification(
  classification_prediction_matrix,
  y,
  minmidmax = "min"
)
}
\arguments{
\item{classification_prediction_matrix}{prediction values}

\item{y}{real values}

\item{minmidmax}{minmidmax}
}
\value{
Q2
}
\description{
Q2 calculation for classification
}
