% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdcvNetParams.R
\name{rdcvNetParams}
\alias{rdcvNetParams}
\title{Make custom parameters for rdcvNet internal modelling
Custom parameters can be set in the function call or by manually setting "slots" in the resulting methParam object}
\usage{
rdcvNetParams(
  robust = 0.05,
  family = "gaussian",
  nRepInner = 1,
  NZV = TRUE,
  oneHot = TRUE
)
}
\arguments{
\item{robust}{Robustness (slack) criterion for determining min and max knees (defaults to 0.05)}

\item{family}{the options could be "gaussian", "binomial", "poisson", "multinomial", "cox", "mgaussian"}

\item{nRepInner}{how many nRepInner}

\item{NZV}{NZV}

\item{oneHot}{TRUE or FALSE using onehot endcoding or not}
}
\value{
a `methParam` object
}
\description{
Make custom parameters for rdcvNet internal modelling
Custom parameters can be set in the function call or by manually setting "slots" in the resulting methParam object
}
\examples{
# Standard parameters for rdcvNet
methParam <- rdcvNetParams()
}
