% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xotu}
\alias{Xotu}
\title{Microbiota composition in mosquitos for the classification tutorial}
\usage{
data(mosquito)
}
\description{
Microbiota composition in mosquitos for the classification tutorial
}
\keyword{data}
