% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XRVIP}
\alias{XRVIP}
\title{Metabolomics data for the rye metabolomics regression tutorial}
\usage{
data(freelive)
}
\description{
Metabolomics data for the rye metabolomics regression tutorial
}
\keyword{data}
