% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XRVIP2}
\alias{XRVIP2}
\title{Metabolomics data for the rye metabolomics regression tutorial, using unique individuals}
\usage{
data(freelive2)
}
\description{
Metabolomics data for the rye metabolomics regression tutorial, using unique individuals
}
\keyword{data}
