% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yotu}
\alias{Yotu}
\title{Village of capture of mosquitos for the classification tutorial}
\usage{
data(mosquito)
}
\description{
Village of capture of mosquitos for the classification tutorial
}
\keyword{data}
