% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YR2}
\alias{YR2}
\title{Rye consumption for the rye metabolomics regression tutorial, using unique individuals}
\usage{
data(freelive2)
}
\description{
Rye consumption for the rye metabolomics regression tutorial, using unique individuals
}
\keyword{data}
